+++ MapboxCoreNavigation/RouteProgress.swift
@@ -388,7 +388,7 @@ open class RouteStepProgress: NSObject {
     /**
      The next intersection the user will travel through.
      
-     The step must contain `Intersections` for this value not be `nil`.
+     The step must contain `intersectionsIncludingUpcomingManeuverIntersection` otherwise this property will be `nil`.
      */
     @objc public var upcomingIntersection: Intersection? {
         guard let intersections = intersectionsIncludingUpcomingManeuverIntersection, intersections.startIndex..<intersections.endIndex-1 ~= intersectionIndex else {
@@ -406,7 +406,7 @@ open class RouteStepProgress: NSObject {
     /**
      The current intersection the user will travel through.
      
-     The step must contain `Intersections` for this value not be `nil`.
+     The step must contain `intersectionsIncludingUpcomingManeuverIntersection` otherwise this property will be `nil`.
      */
     @objc public var currentIntersection: Intersection? {
         guard let intersections = intersectionsIncludingUpcomingManeuverIntersection, intersections.startIndex..<intersections.endIndex ~= intersectionIndex else {
@@ -417,20 +417,9 @@ open class RouteStepProgress: NSObject {
     }
     
     /**
-     Returns all the intersection distances on the current step.
+     Returns an array of the calculated distances from the current intersection to the next intersection on the current step.
      */
     @objc public var intersectionDistances = [CLLocationDistance]()
-    
-    /**
-     The current intersection distance the user will travel through.
-     
-     The step must contain `IntersectionDistances` for this value not to be `nil`.
-     */
-    public var currentIntersectionDistance: CLLocationDistance? {
-        guard intersectionDistances.count > 0 else { return nil }
-        
-        return intersectionDistances[intersectionIndex]
-    }
     
     /**
      The distance in meters the user is to the next intersection they will pass through.