+++ MapboxNavigation/StyleManager.swift
@@ -108,15 +108,8 @@ open class StyleManager: NSObject {
                 delegate?.styleManager?(self, didApply: style)
             }
         }
-        
-        for window in UIApplication.shared.windows {
-            for view in window.subviews {
-                view.removeFromSuperview()
-                window.addSubview(view)
-            }
-        }
-        
-        delegate?.styleManagerDidRefreshAppearance?(self)
+
+        forceRefreshAppearance()
     }
     
     func applyStyle() {
@@ -160,7 +153,10 @@ open class StyleManager: NSObject {
         }
         
         applyStyle()
-        
+        forceRefreshAppearance()
+    }
+    
+    func forceRefreshAppearance() {
         for window in UIApplication.shared.windows {
             for view in window.subviews {
                 view.removeFromSuperview()